basics: a complicated serial blinker
I want to use the SML2 library (header) to represent system states,
manage connections, and the like. Since there's not a user-controllable
LED on the wt32-eth01, set up a hacky "blinky" demo instead using a
serial message. But this acts as a demo of using SML2 to process events.