fix inconsistencies in time update
Makes time update on PPS by incrementing by 1s even if there is no
serial update.  Makes the serial time update responsible for resetting
the time state rather than doing it in the display update. Ensures that
the NTP reference time and uc timestamp are recorded in the same
operation.